requirements.txt
emoji
clean-text
flashtext
ftfy
faiss-cpu
sentence-transformers